+++ src/main/java/com/nephest/battlenet/sc2/model/local/dao/QueueStatsDAO.java
@@ -107,7 +107,7 @@ public class QueueStatsDAO
         );
     }
 
-    public void calculateForSeason(long season)
+    public void calculateForSeason(int season)
     {
         MapSqlParameterSource params = new MapSqlParameterSource()
             .addValue("seasonId", season);
@@ -116,7 +116,7 @@ public class QueueStatsDAO
         LOG.debug("Calculated queue stats for {} season", new Object[]{season});
     }
 
-    public void mergeCalculateForSeason(long season)
+    public void mergeCalculateForSeason(int season)
     {
         MapSqlParameterSource params = new MapSqlParameterSource()
             .addValue("seasonId", season);
