+++ src/main/java/com/nephest/battlenet/sc2/model/local/ladder/dao/LadderStatsDAO.java
@@ -160,7 +160,7 @@ public class LadderStatsDAO
     )
     public Map<Region, Map<BaseLeague.LeagueType, Map<BaseLeagueTier.LeagueTierType, Integer[]>>> findLeagueBounds
     (
-        long season,
+        int season,
         Set<Region> regions,
         Set<League.LeagueType> leagueTypes,
         QueueType queueType,
