+++ src/main/java/com/nephest/battlenet/sc2/model/local/ladder/dao/LadderUtil.java
@@ -54,7 +54,7 @@ public class LadderUtil
 
     public MapSqlParameterSource createSearchParams
     (
-        long season,
+        int season,
         Set<Region> regions,
         Set<League.LeagueType> leagueTypes,
         QueueType queueType,
